+++ packages/themes/scripts/finalize-build.mjs
@@ -10,7 +10,7 @@ const getWorkspaceVersion = async (relativeManifestPath) => {
 };
 
 const manifest = JSON.parse(await readFile(buildPackageUrl, 'utf8'));
-manifest.dependencies['@tale-ui/core'] = await getWorkspaceVersion('../css/package.json');
+manifest.dependencies['@tale-ui/css'] = await getWorkspaceVersion('../css/package.json');
 manifest.dependencies['@tale-ui/utils'] = await getWorkspaceVersion('../utils/package.json');
 
 await writeFile(buildPackageUrl, `${JSON.stringify(manifest, null, 2)}\n`);